The cutover: the in-daemon host dies (#315)
Both chat surfaces route to the overseer container. The bridge's
overseerTurn and the Chat screen's driverFor now call OverseerClient,
and the in-daemon OverseerHost is deleted whole with its module.

The soak door POST /overseer/turn is gone: real operator chat is the
soak now. The boundary comment moved to the head of overseerclient.mjs,
because that file is the daemon's whole reach into the container.
ADR-0014 flips to built. The catalogue tests moved to
overseerverbs.test.mjs, and the conversation-state tests moved into the
boundary suite. The Chat screen stops saying the overseer holds no
shell, and says what the limit is now: a read-only token, and every
effect crossing the daemon.

ADR-0015's replay of a restart-killed turn stays unbuilt and is
surfaced as #395.
