+++ CONTEXT.md
@@ -138,7 +138,7 @@ The command brain of curia. The standing design is one brain with three skins (D
 The container that hosts the overseer. One long-lived service beside the dashboard. It is never a pane and never one container per conversation, because the overseer is not shaped like an agent: it holds many conversations at once and it has no screen. It runs its own image, built from `deploy/overseer/Dockerfile`, which carries git, gh and a reading shell and no build toolchain — so the overseer reads every watched repo and runs no test suite. It is the one service off the host network: a shell in it reaches the daemon at `host.docker.internal` and no other surface on the box, and the daemon reaches it on one published loopback port. Compose owns its liveness, and the daemon only health-checks it. It serves two routes and no third: the health check, and one operator message. See [ADR-0015](docs/adr/0015-the-overseer-is-a-service.md), [#327](https://github.com/alp82/curia/issues/327) and [#314](https://github.com/alp82/curia/issues/314).
 
 **Standing orders**:
-The text the overseer model reads before every turn, plus the tool list that agrees with it. One function writes both, in `daemon/src/overseerprompt.mjs`, and a flag picks the posture: no shell for the in-daemon host, a shell for the overseer service. The shell posture states where the overseer checkouts are, that the checkout pass fetched them once before this turn, and that the overseer token cannot write. It is composed per turn, because the checkout path holds the workspace root and the watch list changes. It STATES the posture and enforces none of it, because a standing order cannot hold a shell. What enforces the posture is four controls: the read-only overseer token, the container, the ✅/❌ confirm on `cancel`, and the tool list. The rest are manners, and a shell undoes each in one command. The text never marks that difference itself, because a line naming an unenforced rule is a line hostile text can quote back. See [ADR-0014](docs/adr/0014-the-overseer-in-its-own-container.md) and [#328](https://github.com/alp82/curia/issues/328).
+The text the overseer model reads before every turn, plus the tool list that agrees with it. One function writes both, in `daemon/src/overseerprompt.mjs`, and a flag picks the posture: a shell for the overseer service, and the no-shell text the retired in-daemon host sent, kept as the pinned record. The shell posture states where the overseer checkouts are, that the checkout pass fetched them once before this turn, and that the overseer token cannot write. It is composed per turn, because the checkout path holds the workspace root and the watch list changes. It STATES the posture and enforces none of it, because a standing order cannot hold a shell. What enforces the posture is four controls: the read-only overseer token, the container, the ✅/❌ confirm on `cancel`, and the tool list. The rest are manners, and a shell undoes each in one command. The text never marks that difference itself, because a line naming an unenforced rule is a line hostile text can quote back. See [ADR-0014](docs/adr/0014-the-overseer-in-its-own-container.md) and [#328](https://github.com/alp82/curia/issues/328).
 _Avoid_: system prompt (that names one of the two things this word covers, and the tool list is the other half).
 
 **Conversation**:
@@ -587,7 +587,7 @@ One box runs everything. Phones and PCs are pure clients on the tailnet.
 - **tmux**: the live agent sessions.
 - **tailscaled**: the Serve rules for attach, timeline, the dashboard, and previews.
 - **Workspace root** (`~/curia-work`): private clones, review checkouts, agent config dirs, and the overseer's checkouts under `overseer/repos/`. Those last are a cache of origin and nothing else, so deleting one costs a re-clone and no work.
-- **Host credential stores** (`~/.claude`, `~/.codex`): the overseer's, which runs on the host and holds no copy. A dispatched agent is in a container and cannot reach them, so it gets the model credential copied into its container environment.
+- **Host credential stores** (`~/.claude`, `~/.codex`): the daemon's own. No container reaches them — a dispatched agent gets the model credential copied into its container environment, and the overseer container reads its own from `daemon/.env.overseer`.
 - **docker**: the live agent containers and the two shared cache volumes.
 
 Everything else is a cache that reconcile can rebuild.
